Move config adjustments earlier
