+++ enviro/constants.py
@@ -47,6 +47,3 @@ UPLOAD_SKIP_FILE = 4
 WATER_VAPOR_SPECIFIC_GAS_CONSTANT = 461.5
 CRITICAL_WATER_TEMPERATURE = 647.096
 CRITICAL_WATER_PRESSURE = 22064000
-
-# default values
-DEFAULT_USB_POWER_TEMPERATURE_OFFSET = 4.5
